move top-level functions from expr to functions
